% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARE.R
\name{ARE}
\alias{ARE}
\title{Factorize of a Tensor}
\usage{
ARE(
  N,
  K,
  r,
  p,
  entity_names,
  link_names,
  attribute_names,
  indirect_link_names,
  lamdaA,
  lamdaR,
  lamdaW,
  X,
  M,
  error_tolerance
)
}
\arguments{
\item{N, K, r, p, entity_names, link_names, attribute_names, indirect_link_names, lamdaA, lamdaR, lamdaW, X, M, error_tolerance}{14 inputs}
}
\value{
A list containing A, the attributes, R, attribute relations and W, the meta-relations
}
\description{
This function factorizes a tensor into the sum of two products. The first term is the product of attributes
and attribute relations, and the second term is the product of the indirect relations and
meta-relations between indirect relations and direct relations.
}
